feat: add GitHub Actions workflow for Manim video rendering
Create a GitHub Actions workflow to automate the rendering of Manim 
videos upon pull requests to the main branch. The workflow includes 
steps for checking out the code, rendering specified scenes from 
`vis.py`, and saving the output as artifacts. This change streamlines 
the video rendering process and ensures consistency in builds.
